Add files via upload
couchdb.py and mcecs_picture_rc.py files for gui prototype. Place them both in the same folder and run couchdb.py. Works on python 3.5. (possibly python 3.7, this is untested as the directions for pyqt5 called for python 3.5) The queries work on python 2.7. It will be easier to convert the queries to python 3.5 to adapt them to work together. Right now, the button clicks do nothing as this is only a prototype!